Added scale model
Added a starter scale model with working balancing. There is currently a jittering bug that causes cubes to easily fall off; I think we might be able to fix it by using a hinge joint instead of the current method of pivoting which is just a fixed rigid body with a hard stop.