99% fixed robot physics
Fixed robot physics to use forces and torques to drive, just needs tuned a bit. We might need to add a way to force it to the ground harder when tipping but still drive fluidly and be able to go up the platforms. We also might need to make the wrist and arm movement be powered by forces and/or torques if clipping persists as a major problem.