Fixed arm making drive base go crazy when raising/lowering, added better decel
Don't really know what I did to solve the first part but whatever it was it worked (I though I reset all changes back to what it was and it still was fixed). Added a constant deceleration on driving and turning, but that still needs tuned. Also readded the scoring camera.